ava_voice_ai/requirements.txt
flet>=0.21.2
requests>=2.31.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
google-generativeai>=0.3.2
python-dotenv>=1.0.0
audioplayer>=0.6
pyttsx3>=2.90
murf
